+++ modules/kubernetes/eks-core/variables.tf
@@ -354,17 +354,17 @@ variable "datadog_config" {
   }
 }
 
-variable "starboard_enabled" {
-  description = "Should Starboard be enabled"
+variable "trivy_enabled" {
+  description = "Should trivy be enabled"
   type        = bool
   default     = false
 }
 
-variable "starboard_config" {
-  description = "Configuration for starboard & trivy"
+variable "trivy_config" {
+  description = "Configuration for trivy-operator & trivy"
   type = object({
-    starboard_role_arn = string
-    trivy_role_arn     = string
+    trivy_operator_role_arn = string
+    trivy_role_arn          = string
   })
 }
 
@@ -394,6 +394,12 @@ variable "node_local_dns_dns_ip" {
 
 variable "node_ttl_enabled" {
   description = "Should Node TTL be enabled"
+  type        = bool
+  default     = true
+}
+
+variable "spegel_enabled" {
+  description = "Should Spegel be enabled"
   type        = bool
   default     = true
 }