+++ modules/kubernetes/prometheus/variables.tf
@@ -145,8 +145,8 @@ variable "grafana_agent_enabled" {
   default     = false
 }
 
-variable "starboard_enabled" {
-  description = "Should starboard be enabled"
+variable "trivy_enabled" {
+  description = "Should trivy be enabled"
   type        = bool
   default     = false
 }
@@ -165,6 +165,12 @@ variable "node_local_dns_enabled" {
 
 variable "promtail_enabled" {
   description = "Should promtail be enabled"
+  type        = bool
+  default     = false
+}
+
+variable "node_ttl_enabled" {
+  description = "Should Node TTL be enabled"
   type        = bool
   default     = false
 }