+++ validation/kubernetes/spegel/main.tf
@@ -1,0 +1,9 @@
+terraform {}
+
+provider "kubernetes" {}
+
+provider "helm" {}
+
+module "spegel" {
+  source = "../../../modules/kubernetes/spegel"
+}